Fix param cost from double to int.

+++ src/org/hsm/controller/ControllerImpl.java
@@ -51,7 +51,7 @@ public class ControllerImpl implements Controller, Serializable {
     }
 
     @Override
-    public void crateGreenhouse(final String name, final GreenHouseType greenhouseType, final double cost, final double size) {
+    public void crateGreenhouse(final String name, final GreenHouseType greenhouseType, final int cost, final double size) {
         this.greenhouse = new GreenhouseImp(name, size, cost);
         this.database = new DBplants();
         this.load = true;
